wrapper/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Dance==7.0.0
clickhouse-driver==0.2.6
docker==7.0.0